Tight Ass Generation Changes Straight from Github
Muuu fukas won;t let me work on shit right in Unity for some reason, so I'll be working a bit straight from GitHub until that's fixed.  I redid all the work I lost this morning, adding a populateMap() function (cycles through all places on the array and Instantiates the correlating object for the place. 

Experimental chance function

Resource Load is attempted, not sure how to use it yet. WILL PROBABLY MAKE KAYLA'S work easier for her, especially if we scrap this project and start a new one since I keep fucking it up.

Enum created for all the ('objects'?) but needs to be cast as an (int) == useful fam.

GameObject player can be called and used to place them in the world. Maybe put that in another script though FAmm

Random chance for map generation will be added/finished later